full coding app-07-04-2024

+++ src/main/java/com/a08r/awcolabsapp/repositores/ILabsDailyTestsRepository.java
@@ -26,7 +26,7 @@ public interface ILabsDailyTestsRepository extends JpaRepository<LabsDailyTestsE
 //    List<LabsDailyTestsEntity> Lab_Parameters(@Param("Lab_Code") int labCode, @Param("Test_Date") String testDate);
 
     @Procedure(name = "Lab_Parameters") //return value equal to Double not (Entity or DTO)
-    List<Double> Lab_Parameters(@Param("Lab_Code") int labCode, @Param("Test_Date") String testDate);
+    List<Float> Lab_Parameters(@Param("Lab_Code") int labCode, @Param("Test_Date") String testDate);
 
 
 }
